Change uint104 to uint256

+++ contracts/tests/TestShareMath.sol
@@ -7,15 +7,15 @@ contract TestShareMath {
     function underlyingToShares(
         uint256 underlyingAmount,
         uint256 pps,
-        uint8 decimals
-    ) external pure returns (uint104) {
+        uint256 decimals
+    ) external pure returns (uint256) {
         return ShareMath.underlyingToShares(underlyingAmount, pps, decimals);
     }
 
     function sharesToUnderlying(
         uint256 shares,
         uint256 pps,
-        uint8 decimals
+        uint256 decimals
     ) external pure returns (uint256) {
         return ShareMath.sharesToUnderlying(shares, pps, decimals);
     }
